Fix a soundness issue with local correctness
Prior to this commit, one local correctness method was generated for each program component. Because assertions are used to denote the Owicki-Gries reasoning, this can jeopardize soundness. In a particular example, Dafny would happily prove a program component containing the Hoare triple ` { true } x := x + 1 { x == 0 || x == 2 }` locally correct, even though this is not a valid Hoare triple.

This commit aims to guarantee soundness by generating a new method for each atomic statement of a program component, with the statement itself as body and the surrounding assertions as preconditions and postconditions.